[API] Enabled delete action not to display error message even if resource already doesn't exist since specs state that Http DELETE verb should be idempotent

+++ src/GGY/ApiBundle/Controller/CategoryController.php
@@ -68,8 +68,11 @@ class CategoryController extends Controller
         $category = $em->getRepository('GGYDataBundle:Category')
             ->find($request->get('id'));
 
-        $em->remove($category);
-        $em->flush();
+        if ($category){
+            $em->remove($category);
+            $em->flush();
+        }
+
     }
 
 }
